+++ ppx/jsonschema/test/shared/cases.ml
@@ -19,29 +19,15 @@ type with_modules = { m : Mod1.m_1; m2 : Mod1.Mod2.m_2 }
 type kind = Success | Error | Skipped [@name "skipped"]
 [@@deriving jsonschema]
 
-type kind_as_string = Success | Error | Skipped [@name "skipped"]
-[@@deriving jsonschema ~variant_as_string]
-
 type poly_kind = [ `Aaa | `Bbb | `Ccc [@name "ccc"] ]
 [@@deriving jsonschema]
-
-type poly_kind_as_string = [ `Aaa | `Bbb | `Ccc [@name "ccc"] ]
-[@@deriving jsonschema ~variant_as_string]
 
 type poly_kind_with_payload =
   [ `Aaa of int | `Bbb | `Ccc of string * bool [@name "ccc"] ]
 [@@deriving jsonschema]
 
-type poly_kind_with_payload_as_string =
-  [ `Aaa of int | `Bbb | `Ccc of string * bool [@name "ccc"] ]
-[@@deriving jsonschema ~variant_as_string]
-
 type poly_inherit = [ `New_one | `Second_one of int | poly_kind ]
 [@@deriving jsonschema]
-
-type poly_inherit_as_string =
-  [ `New_one | `Second_one of int | poly_kind_as_string ]
-[@@deriving jsonschema ~variant_as_string]
 
 type event = {
   date : float;
@@ -105,9 +91,6 @@ type numbers = int list [@@deriving jsonschema]
 type opt = int option [@@deriving jsonschema]
 type using_m = { m : Mod1.m_1 } [@@deriving jsonschema]
 
-type 'param2 poly2 = C of 'param2
-[@@deriving jsonschema ~variant_as_string]
-
 type tuple_with_variant = int * [ `A | `B [@name "second_cstr"] ]
 [@@deriving jsonschema]
 
@@ -140,26 +123,13 @@ type tt = {
 
 type c = char [@@deriving jsonschema]
 
-type variant_inline_record = A of { a : int } | B of { b : string }
-[@@deriving jsonschema ~variant_as_string]
-
 type inline_record_with_extra_fields =
   | User of { name : string; email : string }
       [@jsonschema.allow_extra_fields]
   | Guest of { ip : string }
 [@@deriving jsonschema]
 
-type variant_with_payload =
-  | A of int
-  | B
-  | C of int * string
-  | D of (int * string * bool)
-[@@deriving jsonschema ~variant_as_string]
-
 type t1 = Typ | Class of string [@@deriving jsonschema]
-
-type t2 = Typ | Class of string
-[@@deriving jsonschema ~variant_as_string]
 
 type t3 = Typ [@name "type"] | Class of string [@name "class"]
 [@@deriving jsonschema]
@@ -204,9 +174,6 @@ type 'a param_list = 'a list [@@deriving jsonschema]
 type ('a, 'b) either = Left of 'a | Right of 'b [@@deriving jsonschema]
 type ('a, 'b) either_alias = ('a, 'b) either [@@deriving jsonschema]
 
-type ('a, 'b) direction = North | South
-[@@deriving jsonschema ~variant_as_string]
-
 type tool_params = {
   query : string; [@jsonschema.description "The search query to execute"]
   max_results : int;
@@ -240,11 +207,6 @@ type described_variant =
 type described_variant_inline_record =
   | Point of { x : int; y : int } [@jsonschema.description "A 2D point"]
 [@@deriving jsonschema]
-
-type described_variant_string =
-  | A [@jsonschema.description "First choice"]
-  | B [@jsonschema.description "Second choice"]
-[@@deriving jsonschema ~variant_as_string]
 
 (* The [~ocaml_doc] flag opts into using [(** ... *)] doc comments as a
    fallback for [@jsonschema.description]. Without the flag, doc comments are